+++ elements/CHANGELOG.md
@@ -12,7 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - Can now handle new events of type `pod-os:module`, to asynchronously load a module and pass it to the component that triggered the event. Currently only the `contacts` module is supported.
   - new property `restore-previous-session` allows to restore the session after a refresh or opening in new tab
   - fires new event `pod-os:session-restored` whenever a session is restored
-
+- [pos-router](../docs/elements/components/pos-router)
+  - will navigate back to the original URL after a session has been restored
+  
 ## 0.16.0
 
 ### Added